Squashed commit of the following:
commit 7cbd6da714190b9b0491af19797e135700385fb9
Author: KANG DONGYEONG <50406861+kangddong@users.noreply.github.com>
Date:   Mon Feb 19 18:39:12 2024 +0900

    [WEAV-159] 카카오 로그인 추가 구현 (#15)

    * [MOD] SecretKey add to Gitignore

    * [Chore] Kakao SPM 추가, 프로젝트 설정 변경(기기, 화면방향)

    * [ADD] apple, kakao add image resources

    * [ADD] kakao SDK add to dependency, change Deployment info

    * [Feat] LoginView 작업

    * Merge pull request #11 from Student-Center/infra-applyNetworkSecretKey

    [Infra] Network base url, 보안 적용

    * WEAV-159 [FEAT] KAKAO Login 구현

    * [Feat] LoginView, KakaoLoginButton 분리

    * [FIX] 사용하지않는 isTapped 제거

commit 63669f2d106877d435778bd6b1baa42f75d8ba3f
Author: Jisu Kim <jisu15.kim@gmail.com>
Date:   Mon Feb 19 10:53:46 2024 +0900

    [MOD] Core, DesignSystem 모듈 Deployment Target Version 추가
