+++ weave-iOS/Projects/App/Sources/Login/LoginView.swift
@@ -7,74 +7,28 @@
 
 import SwiftUI
 import DesignSystem
-import KakaoSDKCommon
-import KakaoSDKAuth
-import KakaoSDKUser
 import Services
 
 struct LoginView: View {
     var body: some View {
         VStack {
             Spacer()
+            
             DesignSystem.Icons.appLogo
-            
             Spacer()
-            Button(action: {
-                isAvailableOpenKakao()
-            }, label: {
-                HStack {
-                    DesignSystem.Icons.kakaoLogo
-                    Text("카카오 로그인")
-                        .foregroundColor(.black)
-                }
-                
-            })
-            .padding(20)
-            .frame(width: 300, height: 44, alignment: .center)
-            .background(Color(red: 1, green: 0.9, blue: 0))
-            .cornerRadius(6)
             
+            KakaoLoginButton(onComplte: { idToken in
+                Task {
+                    await requestSNSLogin(idToken: idToken, with: .kakao)
+                }
+            })
             Spacer()
                 .frame(height: 16)
             
-            Button(action: {
-                
-            }, label: {
-                HStack {
-                    DesignSystem.Icons.appleLogo
-                    Text("Apple로 로그인")
-                        .foregroundColor(.black)
-                }
-                
-            })
-            .padding(.horizontal, 16)
-            .padding(.vertical, 12)
-            .frame(width: 300, height: 44, alignment: .center)
-            .background(.white)
-            .cornerRadius(6)
+//            AppleLoginButton()
+            
             Spacer()
                 .frame(height: 58)
-        }
-    }
-    
-    private func isAvailableOpenKakao() {
-        if UserApi.isKakaoTalkLoginAvailable() {
-            UserApi.shared.loginWithKakaoTalk { oauthToken, error in
-                if let error = error {
-                    print(error)
-                } else {
-                    print("loginWithKakaoTalk() success.")
-                    if let idToken = oauthToken?.idToken {
-                        print("oauthToken: \(idToken)")
-                        Task {
-                            await requestSNSLogin(idToken: idToken, with: .kakao)
-                            // 화면이동처리
-                        }
-                    }
-                    
-                    _ = oauthToken
-                }
-            }
         }
     }
